Split AttachmentStoreStrategy classes to new FileStoreStrategy classes
